User Experience/User Inferface completed updated. Changes from Web3 Enabler v1.4.1 incorporated.
The Field Labels were updated to utilize XRPLedger names.
The Field (API) Names were preserved to maintain code compatibility.

Networks was removed to be consistent with Web3 Enabler behavior.

Flows have been migrated but need to be rebuild to refllect the different worldview of the XRP Ledger from the EVM Model.