Initial release of Milestone 1.
Building upon the initial demo, this initial commitment includes the basic Salesforce UX for XRP,
plus the supporting APEX Calls.

This is a fully interactive proof of concept, allowing us to proceed to the next step. We will focus
on implementing the business logic, and creating both in-app and offline documentation.